final_project/requirements.txt
certifi==2020.12.5
chardet==4.0.0
decorator==4.4.2
future==0.18.2
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.3
Js2Py==0.71
moviepy==1.0.3
mutagen==1.45.1
numpy==1.20.2
Pillow==8.2.0
proglog==0.1.9
pygame==2.0.1
pyjsparser==2.7.1
pytube==10.7.1
pytz==2021.1
requests==2.25.1
six==1.15.0
tqdm==4.60.0
tube-dl==5.1.0
tzlocal==2.1
urllib3==1.26.4
